Refactoring Action and MonitoredList out of core.py
